[#254] Fix : 홈 비즈니스 로직 코드 정리
