+++ presentation/home/src/main/java/com/dhc/home/model/HomeUiModel.kt
@@ -48,69 +48,3 @@ data class HomeUiModel(
         )
     }
 }
-
-data class MissionUiModel(
-    val missionId: String = "",
-    val category: String = "",
-    val type: MissionType = MissionType.LONG_TERM,
-    val finished: Boolean = false,
-    val title: String = "",
-    val endDate: String = "",
-    val difficulty: String = "",
-    val switchCount: Int = 0,
-    val isBlink: Boolean = false,
-    val isExpanded: Boolean = false
-) {
-    companion object {
-        fun Int.calDifficulty(): String {
-            return when (this) {
-                1 -> "Easy"
-                2 -> "Medium"
-                3 -> "Hard"
-                else -> "Easy"
-            }
-        }
-
-        fun String.toDDay(): String {
-            return try {
-                val formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd")
-                val targetDate = LocalDate.parse(this, formatter)
-                val today = LocalDate.now()
-
-                val daysBetween = ChronoUnit.DAYS.between(today, targetDate)
-
-                when {
-                    daysBetween > 0 -> "D-${daysBetween}"
-                    daysBetween == 0L -> "D-Day"
-                    else -> "D+${abs(daysBetween)}"
-                }
-            } catch (e: Exception) {
-                "Invalid Date"
-            }
-        }
-    }
-}
-
-fun Mission.toUiModel() = MissionUiModel(
-    missionId = missionId,
-    category = category,
-    type = type,
-    finished = finished,
-    title = title,
-    endDate = endDate.toDDay(),
-    difficulty = difficulty.calDifficulty(),
-    switchCount = switchCount,
-)
-
-data class TodayDailyFortuneUiModel(
-    val date: String = "",
-    val fortuneTitle: String = "",
-    val fortuneDetail: String = "",
-    val score: Int = 0,
-)
-
-
-fun getMissionIdList(longTermMission: MissionUiModel, todayDailyMissionList: List<MissionUiModel>): List<String> {
-    return listOf(longTermMission.missionId) + todayDailyMissionList.map { it.missionId }
-}
-
