+++ presentation/home/src/main/java/com/dhc/home/model/TodayDailyFortuneUiModel.kt
@@ -1,0 +1,8 @@
+package com.dhc.home.model
+
+data class TodayDailyFortuneUiModel(
+    val date: String = "",
+    val fortuneTitle: String = "",
+    val fortuneDetail: String = "",
+    val score: Int = 0,
+)
